+++ src/main/java/com/management/company/controller/CompanyController.java
@@ -2,13 +2,13 @@ package com.management.company.controller;
 
 import com.management.person.dto.PersonDTO;
 import com.management.person.service.PersonService;
+import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import io.swagger.v3.oas.annotations.Operation;
 
 import java.util.List;
 
@@ -25,7 +25,7 @@ public class CompanyController {
 
     @GetMapping("/{companyId}/persons")
     @Operation(summary = "Finds all persons for a company")
-    public ResponseEntity<List<PersonDTO>> findPersonsByCompanyId(@PathVariable Long companyId) {
+    public ResponseEntity<List<PersonDTO>> findPersonsByCompanyId(@PathVariable Integer companyId) {
         return ResponseEntity.ok(personService.findPersonsByCompanyId(companyId));
     }
 }
