+++ src/main/java/com/management/projectoccurrence/controller/ProjectOccurrenceController.java
@@ -4,12 +4,12 @@ import com.management.projectoccurrence.dto.ProjectOccurrenceCreateRequest;
 import com.management.projectoccurrence.dto.ProjectOccurrenceResponse;
 import com.management.projectoccurrence.dto.ProjectOccurrenceUpdateRequest;
 import com.management.projectoccurrence.service.ProjectOccurrenceService;
+import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import io.swagger.v3.oas.annotations.Operation;
 
 import java.util.List;
 
@@ -32,25 +32,25 @@ public class ProjectOccurrenceController {
 
     @GetMapping("/occurrences/{id}")
     @Operation(summary = "Finds an occurrence by ID")
-    public ResponseEntity<ProjectOccurrenceResponse> getById(@PathVariable Long id) {
+    public ResponseEntity<ProjectOccurrenceResponse> getById(@PathVariable Integer id) {
         return ResponseEntity.ok(projectOccurrenceService.findById(id));
     }
 
     @GetMapping("/projects/{projectId}/occurrences")
     @Operation(summary = "Finds all occurrences for a project")
-    public ResponseEntity<List<ProjectOccurrenceResponse>> listByProject(@PathVariable Long projectId) {
+    public ResponseEntity<List<ProjectOccurrenceResponse>> listByProject(@PathVariable Integer projectId) {
         return ResponseEntity.ok(projectOccurrenceService.listByProject(projectId));
     }
 
     @PutMapping("/occurrences/{id}")
     @Operation(summary = "Updates an occurrence")
-    public ResponseEntity<ProjectOccurrenceResponse> update(@PathVariable Long id, @Valid @RequestBody ProjectOccurrenceUpdateRequest request) {
+    public ResponseEntity<ProjectOccurrenceResponse> update(@PathVariable Integer id, @Valid @RequestBody ProjectOccurrenceUpdateRequest request) {
         return ResponseEntity.ok(projectOccurrenceService.update(id, request));
     }
 
     @DeleteMapping("/occurrences/{id}")
     @Operation(summary = "Deletes an occurrence")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
         projectOccurrenceService.delete(id);
         return ResponseEntity.noContent().build();
     }
